+++ src/oci/base_client.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-# Copyright (c) 2016, 2024, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2025, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 
 from __future__ import absolute_import
@@ -703,8 +703,9 @@ class BaseClient(object):
             timestamp = datetime.now(timezone.utc).isoformat()
 
             service_code, message, deserialized_data = self.get_deserialized_service_code_and_message(response, allow_control_chars)
-            if response.status_code == 413 and service_code == 'RequestEntityTooLarge':
-                self.logger.warning("Recieved a 413/RequestEntityTooLarge from {}, resetting session".format(target_service))
+            if (response.status_code == 413 and service_code == 'MissingHeaderBodySeparatorDefect') or (response.status_code == 412 and service_code == 'IfNoneMatchFailed'):
+                self.logger.warning(
+                    f"Received a {response.status_code}/{service_code} from {target_service}, resetting session")
                 _ = response.content  # Read the response content to enable closing the socket.
                 response.close()
                 new_session = copy.copy(self.session)